CACreateCert: improve features for use with --email
Mention in the README that CACreateCert can create perfectly usable
S/MIME email signing/encryption certificates.  Also mention that an
example of how to do so is now part of the `CACreateCert --help`
output.

Also make it possible to move the location of the emailAddress
component of the subject distinguished name to anywhere in the
subject distinguished name if desired.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>