CACreateCert: support --request and --utf8
With --request output a Certificate Signing Request instead of an
X.509 certificate.

With --utf8 always use UTF8String to encode strings that would
otherwise be encoded as PrintableString and are allowed to alternatively
be encoded as UTF8String.

Note that it may be necessary to use both the --no-extensions and
--utf8 options to generate a byte-exact match for `openssl req -new`
output.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>